requirements.txt
alembic==1.6.5
appdirs==1.4.4
bcrypt==3.2.0
blinker==1.4
bokeh==2.3.3
cachelib==0.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
chromedriver-binary==93.0.4577.15.0
click==8.0.1
cmake==3.21.1.post1
config==0.5.0.post0
cryptography==3.4.8
cycler==0.10.0
distlib==0.3.2
dnspython==1.16.0
dominate==2.6.0
email-validator==1.1.3
env==0.1.0
et-xmlfile==1.1.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-Ext==0.1
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-PyMongo==2.3.0
Flask-Scss==0.5
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
openpyxl==3.0.7
packaging==21.0
pafy==0.5.5
pandas==1.3.0
passlib==1.7.4
pathlib==1.0.1
Pillow==8.3.1
pycparser==2.20
pygame==2.0.1
pymongo==3.11.4
pymssql==2.2.2
pyparsing==2.4.7
pyScss==1.3.7
python-dateutil==2.8.1
python-dotenv==0.19.0
python-editor==1.0.4
python-vlc==3.0.12118
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.4.20
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==20.4.7
visitor==0.1.3
Werkzeug==2.0.1
WTForm==1.0
WTForms==2.3.3
xlrd==2.0.1
